fix some stuff

+++ bookstore-frontend/src/store/CategoryStore.ts
@@ -1,13 +1,14 @@
 import { Category } from '@/types/common';
 import { create } from 'zustand';
+import useAuthStore from './AuthStore';
+import { toast } from "react-hot-toast";
 
 interface CategoryState {
     categories: Category[];
     fetchCategories: () => Promise<void>;
-    createCategory: (name: string) => Promise<void>;
+    createCategory: (name: string) => Promise<boolean>;
     updateCategory: (id: number, name: string) => Promise<void>;
     deleteCategory: (id: number) => Promise<void>;
-    loadCategories: (search?:string) => Promise<void>;
 }
 
 const useCategoryStore = create<CategoryState>((set, get) => ({
@@ -26,26 +27,42 @@ const useCategoryStore = create<CategoryState>((set, get) => ({
     },
     createCategory: async(name) => {
         try {
+            const token = useAuthStore.getState().accessToken;
+            if (!token) throw new Error("Unauthorized");
             const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/categories`, {
                 method: 'POST',
                 headers: {
                     'Content-Type' : 'application/json',
+                    Authorization: `Bearer ${token}`,
                 },
                 body: JSON.stringify({ name: name })
             })
+            const responseData = await response.json();
             if(!response.ok){
-                throw new Error("Failed to create data.")
+                if (responseData.errors?.name) { 
+                    toast.error("Validation Error: " + responseData.errors.name[0]);
+                     
+                } else {
+                    toast.error("Failed to create category: " + responseData.message);
+                }
+                toast
+                return false;
             }
+            return true
         } catch (error){
-            throw error;
+            console.error("Error creating category:", error)
+            return false
         }
     },
     updateCategory: async(id, name) => {
         try {
+            const token = useAuthStore.getState().accessToken;
+            if (!token) throw new Error("Unauthorized");
             const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/categories/${id}`, {
                 method: 'PUT',
                 headers: {
                     'Content-Type' : 'application/json',
+                    Authorization: `Bearer ${token}`,
                 },
                 body: JSON.stringify({ name: name })
             })
@@ -59,8 +76,13 @@ const useCategoryStore = create<CategoryState>((set, get) => ({
     },
     deleteCategory: async (id) => {
         try{
+            const token = useAuthStore.getState().accessToken;
+            if (!token) throw new Error("Unauthorized");
             const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/categories/${id}`, {
-                method: "DELETE"
+                method: "DELETE",
+                headers: {
+                    Authorization: `Bearer ${token}`,
+                }
             })
             if(!response.ok){
                 throw new Error("Error in fetch data")
@@ -68,9 +90,6 @@ const useCategoryStore = create<CategoryState>((set, get) => ({
         } catch (error){
             throw error;
         }
-    },
-    loadCategories: async (search="") => {
-
     }
 }))
